Add self field in AbstractCrudService

+++ src/main/java/software/plusminus/data/service/AbstractCrudService.java
@@ -1,10 +1,10 @@
 package software.plusminus.data.service;
 
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
+import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.transaction.support.TransactionSynchronizationManager;
 import software.plusminus.crud.listener.CrudListenerContext;
 import software.plusminus.data.exception.NotFoundException;
 import software.plusminus.data.model.Update;
@@ -17,21 +17,35 @@ import javax.annotation.Nullable;
 import javax.validation.Validator;
 
 @SuppressWarnings("java:S119")
-@NoArgsConstructor
-@AllArgsConstructor
-public abstract class AbstractCrudService<T, ID> implements CrudService<T, ID>  {
+public abstract class AbstractCrudService<T, ID> implements CrudService<T, ID> {
 
     private Validator validator;
     private PatchService patchService;
     private CrudListenerContext listenerContext;
     private CrudRepository<T, ID> repository;
+    @Nullable
+    private AbstractCrudService<T, ID> self;
+
+    protected AbstractCrudService() {
+    }
+
+    protected AbstractCrudService(Validator validator,
+                                  PatchService patchService,
+                                  CrudListenerContext listenerContext,
+                                  CrudRepository<T, ID> repository) {
+        this.validator = validator;
+        this.patchService = patchService;
+        this.listenerContext = listenerContext;
+        this.repository = repository;
+    }
 
     @Autowired
     void init(Validator validator,
               PatchService patchService,
               CrudListenerContext listenerContext,
               @Nullable CrudRepository<T, ID> repository,
-              RepositoryContext repositoryContext) {
+              RepositoryContext repositoryContext,
+              ObjectProvider<AbstractCrudService<T, ID>> self) {
         if (this.validator == null) {
             this.validator = validator;
         }
@@ -44,10 +58,14 @@ public abstract class AbstractCrudService<T, ID> implements CrudService<T, ID>  
         if (this.repository == null) {
             this.repository = DataUtil.provideCrudRepository(repository, repositoryContext, this, CrudService.class);
         }
+        this.self = self.getIfUnique();
     }
 
     @Override
     public T getById(ID id) {
+        if (self != null && !TransactionSynchronizationManager.isActualTransactionActive()) {
+            return self.getById(id);
+        }
         T object = repository.getById(id);
         if (object == null) {
             throw new NotFoundException("Can't find object with id " + id);
@@ -58,6 +76,9 @@ public abstract class AbstractCrudService<T, ID> implements CrudService<T, ID>  
 
     @Override
     public Page<T> getPage(Pageable pageable) {
+        if (self != null && !TransactionSynchronizationManager.isActualTransactionActive()) {
+            return self.getPage(pageable);
+        }
         Page<T> page = repository.findAll(pageable);
         page.forEach(listenerContext::onRead);
         return page;
@@ -65,6 +86,9 @@ public abstract class AbstractCrudService<T, ID> implements CrudService<T, ID>  
 
     @Override
     public T create(T object) {
+        if (self != null && !TransactionSynchronizationManager.isActualTransactionActive()) {
+            return self.create(object);
+        }
         DataUtil.verifyOnCreate(object);
         listenerContext.beforeCreate(object);
         T created = repository.save(object);
@@ -74,6 +98,9 @@ public abstract class AbstractCrudService<T, ID> implements CrudService<T, ID>  
 
     @Override
     public T update(T object) {
+        if (self != null && !TransactionSynchronizationManager.isActualTransactionActive()) {
+            return self.update(object);
+        }
         DataUtil.verifyOnUpdate(object);
         listenerContext.beforeUpdate(object);
         T updated = repository.save(object);
@@ -83,6 +110,9 @@ public abstract class AbstractCrudService<T, ID> implements CrudService<T, ID>  
 
     @Override
     public T patch(T patch) {
+        if (self != null && !TransactionSynchronizationManager.isActualTransactionActive()) {
+            return self.patch(patch);
+        }
         ID id = DataUtil.verifyOnPatch(patch);
         T target = getById(id);
         listenerContext.beforePatch(patch);
@@ -95,9 +125,21 @@ public abstract class AbstractCrudService<T, ID> implements CrudService<T, ID>  
 
     @Override
     public void delete(T object) {
+        if (self != null && !TransactionSynchronizationManager.isActualTransactionActive()) {
+            self.delete(object);
+            return;
+        }
         DataUtil.verifyOnDelete(object);
         listenerContext.beforeDelete(object);
         repository.delete(object);
         listenerContext.afterDelete(object);
+    }
+
+    @Override
+    public T save(T object) {
+        if (self != null && !TransactionSynchronizationManager.isActualTransactionActive()) {
+            return self.save(object);
+        }
+        return CrudService.super.save(object);
     }
 }
